# api/openapi.yaml
openapi: 3.1.0
info:
  title: Mixologist API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    HTTP-контракт Mixologist. Источник истины для request/response DTO и error shapes.

    Внешний API версионирован через префикс `/api/v1`. Operational health (`GET /health`)
    остаётся вне продуктового контракта.

    Авторизация — Google OAuth на backend, opaque HttpOnly cookie-сессия.
    Mutating API с cookie-сессией защищается CSRF-заголовком `X-CSRF-Token`.
    Чтение ingredients публичное; мутации ingredients — только `admin`.
servers:
  - url: /
    description: Same-origin (Vite/nginx proxy или backend)
security:
  # CookieSession опционален: `{}` = без auth (публичные GET).
  - CookieSession: []
  - {}
...
paths:
...
  /api/v1/ingredients:
    get:
      operationId: listIngredients
      tags: [ingredients]
      summary: Список ингредиентов
      description: |
        Token-based (keyset) пагинация. Публичные имена параметров — `pageToken`/`pageSize`.

        Default sort: `created_at DESC` (с внутренним tie-breaker `id DESC`).
        В одном запросе допускается только одно поле сортировки.
        `totalSize` — точное число записей после фильтров, до ограничения текущей страницы.
        Пустая строка `nextPageToken` означает отсутствие следующей страницы.

        Публичный endpoint: аутентификация не требуется.
      security: []
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/IngredientSort"
        - $ref: "#/components/parameters/SortOrder"
        - $ref: "#/components/parameters/IngredientNameFilter"
        - $ref: "#/components/parameters/IngredientTypeFilter"
        - $ref: "#/components/parameters/AbvFilter"
      responses:
        "200":
          description: Страница ингредиентов
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngredientListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    post:
      operationId: createIngredient
      tags: [ingredients]
      summary: Создать ингредиент
      description: |
        Права: только `admin`.
      security:
        - CookieSession: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngredientRequest"
      responses:
        "201":
          description: Ингредиент создан
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/ingredients/{id}:
    parameters:
      - $ref: "#/components/parameters/IngredientID"
    get:
      operationId: getIngredient
      tags: [ingredients]
      summary: Получить ингредиент по ID
      description: |
        Публичный endpoint: аутентификация не требуется.
      security: []
      responses:
        "200":
          description: Ингредиент найден
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    patch:
      operationId: updateIngredient
      tags: [ingredients]
      summary: Частично обновить ингредиент
      description: |
        Optimistic locking: поле `version` обязательно и должно совпадать с текущей версией.
        Необходимо передать хотя бы одно изменяемое поле помимо `version`.
        Пустой body `{}` или только `{"version": N}` — `VALIDATION_ERROR`.

        Права: только `admin`.
      security:
        - CookieSession: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIngredientRequest"
      responses:
        "200":
          description: Ингредиент обновлён
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    delete:
      operationId: deleteIngredient
      tags: [ingredients]
      summary: Удалить ингредиент
      description: |
        При использовании ингредиента в рецептах возвращается `409 RESOURCE_IN_USE`
        (реальные FK появятся вместе с recipes).

        Права: только `admin`.
      security:
        - CookieSession: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204":
          description: Ингредиент удалён
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/ingredients/{id}/icon:
    parameters:
      - $ref: "#/components/parameters/IngredientID"
    get:
      operationId: getIngredientIcon
      tags: [ingredients]
      summary: Получить иконку ингредиента
      description: |
        Публичный endpoint: аутентификация не требуется.
      security: []
      responses:
        "200":
          description: Бинарная иконка (PNG или JPEG)
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Content-Length:
              description: Размер тела ответа в байтах
              schema:
                type: integer
                minimum: 1
            Cache-Control:
              description: Кеширование иконки
              schema:
                type: string
                example: public, max-age=300, must-revalidate
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    put:
      operationId: putIngredientIcon
      tags: [ingredients]
      summary: Загрузить иконку ингредиента
      description: |
        Сырое тело файла (PNG или JPEG), максимум 512 KB.
        Права: только `admin`.
      security:
        - CookieSession: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              maxLength: 524288
      responses:
        "204":
          description: Иконка сохранена
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  securitySchemes:
    CookieSession:
      type: apiKey
      in: cookie
      name: session
      description: |
        Opaque HttpOnly cookie-сессия после Google OAuth.
        Чтение ingredients публичное; мутации ingredients — только `admin`.
        Mutating запросы (`POST`, `PUT`, `PATCH`, `DELETE`) с cookie-сессией
        дополнительно требуют CSRF-заголовок `X-CSRF-Token`. Токен выпускается
        сервером (HMAC от session token и часового timestamp), отдаётся клиенту
        в CSRF cookie и проверяется по заголовку в окне ±1 час.
  headers:
    XRequestID:
      description: |
        Корреляционный ID запроса. Принимается из входящего `X-Request-ID`
        (если валиден) либо генерируется сервером. Тот же ID в `error.request_id`.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: |
        CSRF token для mutating запросов с cookie-сессией.
        Выпускается сервером (HMAC от session token и часового timestamp),
        отдаётся в CSRF cookie для чтения JS и проверяется по заголовку.
        Токен принимается в окне ±1 час; cookie обновляется на аутентифицированных запросах.
      schema:
        type: string
        minLength: 1
    IngredientID:
      name: id
      in: path
      required: true
      description: Идентификатор ингредиента (положительное целое)
      schema:
        type: integer
        minimum: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Размер страницы (1–100). Default — 25.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    PageToken:
      name: pageToken
      in: query
      required: false
      description: |
        Opaque base64url token следующей страницы.
        Отвергается (`INVALID_PAGE_TOKEN`), если формат некорректен или
        sort/order/pageSize не совпадают с параметрами текущего запроса.
      schema:
        type: string
        minLength: 1
    IngredientSort:
      name: sort
      in: query
      required: false
      description: Поле сортировки. Допускается только одно значение.
      schema:
        allOf:
          - $ref: "#/components/schemas/IngredientSortField"
        default: created_at
    SortOrder:
      name: order
      in: query
      required: false
      description: |
        Направление сортировки. Для запроса без параметров используется `desc`
        (то есть `created_at DESC`).
      schema:
        allOf:
          - $ref: "#/components/schemas/SortOrder"
        default: desc
    IngredientNameFilter:
      name: name
      in: query
      required: false
      description: Регистронезависимый поиск по подстроке (`ILIKE`)
      schema:
        type: string
        minLength: 1
        maxLength: 512
    IngredientTypeFilter:
      name: ingredient_type
      in: query
      required: false
      description: Точное совпадение enum
      schema:
        $ref: "#/components/schemas/IngredientType"
    AbvFilter:
      name: abv
      in: query
      required: false
      description: Точное совпадение enum
      schema:
        $ref: "#/components/schemas/Abv"
  responses:
    BadRequest:
      description: |
        Ошибка валидации запроса. Типичные коды:
        `VALIDATION_ERROR`, `INVALID_ID`, `INVALID_PAGE_TOKEN`.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              summary: VALIDATION_ERROR
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Некорректные данные запроса
                  request_id: 550e8400-e29b-41d4-a716-446655440000
                  details:
                    - field: name
                      message: длина имени должна быть от 3 до 512 символов
            invalidId:
              summary: INVALID_ID
              value:
                error:
                  code: INVALID_ID
                  message: Неверный ID ингредиента
                  request_id: 550e8400-e29b-41d4-a716-446655440000
            invalidPageToken:
              summary: INVALID_PAGE_TOKEN
              value:
                error:
                  code: INVALID_PAGE_TOKEN
                  message: Некорректный или несовместимый pageToken
                  request_id: 550e8400-e29b-41d4-a716-446655440000
    Unauthorized:
      description: Требуется аутентификация (`UNAUTHORIZED`).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Требуется аутентификация
              request_id: 550e8400-e29b-41d4-a716-446655440000
    Forbidden:
      description: Недостаточно прав (`FORBIDDEN`).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: Недостаточно прав
              request_id: 550e8400-e29b-41d4-a716-446655440000
    NotFound:
      description: Ресурс не найден (`NOT_FOUND`)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: Ингредиент не найден
              request_id: 550e8400-e29b-41d4-a716-446655440000
    Conflict:
      description: |
        Конфликт состояния. Типичные коды:
        `ALREADY_EXISTS`, `VERSION_CONFLICT`, `RESOURCE_IN_USE`.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            alreadyExists:
              summary: ALREADY_EXISTS
              value:
                error:
                  code: ALREADY_EXISTS
                  message: Ингредиент с таким именем уже существует
                  request_id: 550e8400-e29b-41d4-a716-446655440000
            versionConflict:
              summary: VERSION_CONFLICT
              value:
                error:
                  code: VERSION_CONFLICT
                  message: Ингредиент был изменён другим запросом
                  request_id: 550e8400-e29b-41d4-a716-446655440000
            resourceInUse:
              summary: RESOURCE_IN_USE
              value:
                error:
                  code: RESOURCE_IN_USE
                  message: Ингредиент используется и не может быть удалён
                  request_id: 550e8400-e29b-41d4-a716-446655440000
    InternalError:
      description: Внутренняя ошибка (`INTERNAL_ERROR`). Детали БД/stack наружу не отдаются.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: Внутренняя ошибка сервера
              request_id: 550e8400-e29b-41d4-a716-446655440000
    ServiceUnavailable:
      description: Сервис временно недоступен (`SERVICE_UNAVAILABLE`)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: SERVICE_UNAVAILABLE
              message: Сервис временно недоступен
              request_id: 550e8400-e29b-41d4-a716-446655440000
  schemas:
...
    UnitMeasurement:
      type: string
      enum: [мл, гр, шт, дэш]
    Abv:
      type: string
      enum: [безалкогольный, слабоалкогольный, крепкий]
    IngredientType:
      type: string
      enum:
        - крепкая часть
        - безалкогольная часть
        - вермут
        - вино
        - ликер
        - биттер
        - сироп
        - другое
        - фрукт
        - овощ
        - ягода
    IngredientSortField:
      type: string
      enum: [created_at, name]
    SortOrder:
      type: string
      enum: [asc, desc]
    ErrorCode:
      type: string
      description: Стабильный машинный код ошибки (uppercase)
      enum:
        - VALIDATION_ERROR
        - INVALID_ID
        - INVALID_PAGE_TOKEN
        - NOT_FOUND
        - ALREADY_EXISTS
        - VERSION_CONFLICT
        - RESOURCE_IN_USE
        - UNAUTHORIZED
        - FORBIDDEN
        - OAUTH_STATE_INVALID
        - OAUTH_CALLBACK_FAILED
        - CSRF_TOKEN_INVALID
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
          description: Имя поля (JSON path или query param)
          minLength: 1
        message:
          type: string
          description: Безопасное русское сообщение по полю
          minLength: 1
    ErrorBody:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Безопасное русское сообщение для клиента
          minLength: 1
        request_id:
          type: string
          description: Тот же ID, что в заголовке `X-Request-ID`
          minLength: 1
          maxLength: 128
        details:
          type: array
          description: Опциональные field-level ошибки
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    Ingredient:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - description
        - unit_measurement
        - abv
        - ingredient_type
        - has_icon
        - version
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 3
          maxLength: 512
        description:
          type: string
          maxLength: 1024
        unit_measurement:
          $ref: "#/components/schemas/UnitMeasurement"
        abv:
          $ref: "#/components/schemas/Abv"
        ingredient_type:
          $ref: "#/components/schemas/IngredientType"
        has_icon:
          type: boolean
          description: true, если у ингредиента загружена иконка
        version:
          type: integer
          minimum: 1
          description: Версия для optimistic locking
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    IngredientListResponse:
      type: object
      additionalProperties: false
      required: [ingredients, nextPageToken, totalSize]
      properties:
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        nextPageToken:
          type: string
          description: |
            Opaque token следующей страницы. Пустая строка на последней странице.
        totalSize:
          type: integer
          minimum: 0
          description: |
            Точное количество записей после фильтров, независимо от pageSize/pageToken.
    CreateIngredientRequest:
      type: object
      additionalProperties: false
      required: [name, unit_measurement, abv, ingredient_type]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 512
        description:
          type: string
          maxLength: 1024
          default: ""
        unit_measurement:
          $ref: "#/components/schemas/UnitMeasurement"
        abv:
          $ref: "#/components/schemas/Abv"
        ingredient_type:
          $ref: "#/components/schemas/IngredientType"
    UpdateIngredientRequest:
      type: object
      additionalProperties: false
      required: [version]
      minProperties: 2
      description: |
        `version` обязателен. Кроме него нужно хотя бы одно изменяемое поле
        (`name`, `description`, `unit_measurement`, `abv`, `ingredient_type`).
      properties:
        version:
          type: integer
          minimum: 1
          description: Ожидаемая текущая версия записи
        name:
          type: string
          minLength: 3
          maxLength: 512
        description:
          type: string
          maxLength: 1024
        unit_measurement:
          $ref: "#/components/schemas/UnitMeasurement"
        abv:
          $ref: "#/components/schemas/Abv"
        ingredient_type:
          $ref: "#/components/schemas/IngredientType"